requirements.txt
colorama==0.4.5
numpy==1.24.0
Pillow==9.4.0
